% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choroMap.R
\name{choroMap}
\alias{choroMap}
\title{Create choropleth map}
\usage{
choroMap(ds, c, t, p)
}
\arguments{
\item{ds}{dataset}

\item{c}{column}

\item{t}{title}

\item{p}{palette}
}
\value{
a map
}
\description{
Create a choropleth map from a given dataset using
the coloumn of interest
}
