% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Climate.R
\docType{data}
\name{climate}
\alias{climate}
\title{Climate data for Chicago}
\format{
A data frame with 14990 rows and 30 variables:
\describe{
\item{Station}{station, station that collected the data}
\item{Latitude}{latitude, part of the coordinate}
...
}
}
\source{
\url(https://www.ncdc.noaa.gov/cdo-web/)
}
\usage{
climate
}
\description{
A data set containing data related to climate for Chicago
}
\keyword{datasets}
